Look in extended description.
Added error and loading icons, and changed the default icon to loading
in the about box, now I'm waiting for an update system.
When send is pressed, and the user isn't logged in, the login box will
now show, and save afterwards.
Login isn't functional yet as the backend isn't ready.